Migration
Strangely the documented Airtable API doesn't currently allow you to export the schema for your base. But there is a hack that will use the undocumented API that powers the API documentation page to persist a base schema as JSON:

https://gist.github.com/douglaslassance/0ba26f2cf2aa9bb21a521ba07d751244

I put the schemas in the schemas subdirectory, and moved the
documentation pieces into the docs subdirectory. I also started on an
experimental migration program to move data from one schema to the
other.