LDA migration
This commit sets things up to start the LDA data migration. I pulled
some more things out of ldt_tracking.py into schema. I also added a
little viewer for the original file names that are present in the files
that are in the database to help give a sense of what accession records
to create.